feat(core): display refresh rate scaling
In order to lower the power consumption in standard operation mode, the goal is to regulate the display refresh rate based on the current needs.

Since the  display visible dimensions are 520x380 and display controller don't support thisresolution resolution directly, the closest variant 520x480 is used with 50px offset. Changing the resolution used to 520x430 and shortening the HACT region by 50px effectivly increasing the blanking period.

[no changelog]